pangu-l: modify one key power on logic
when system power off:
1. EC need turn off fingerprint keyboard USB port
power and then turn on it after 200ms. This is to
reset the fingerprint module.
2. EC need pull down F460_PB0 and F460_PB1. This is
to inform the HC32F460 of the power state.
3. EC need pull down EC_TO_USB_SWITCH. This is for
switch USB signal to HC32F460.

when system power on:
1. EC need pull up F46_PB0 and F460_PB1. This is to
inform the HC32F460 of the power state.
2. EC need pull up EC_TO_USB_SWITCH after 600ms. This
is for switch USB signal to CPU.

BUG=EC can't power on
BRANCH=pangul-dev-branch
TEST=1. Compile successfully.
     2. one key power on verify pass

Change-Id: I0ba805fd917f8de7d18ffc6fe5f46640567d87e2